fix: Integrate viewModel to ListScreen
refactor List by rename to ListScreen for coherence
